intraday_strategies/0dte_backtesting/requirements.txt
# 0DTE SPY Options Trading System - Requirements

# Core Trading Dependencies
alpaca-py>=0.42.0          # Alpaca trading API
py_vollib>=1.0.1           # Black-Scholes Greeks calculation
python-dotenv>=1.0.0       # Environment variable management
pytz>=2024.1               # Timezone handling
requests>=2.32.0           # HTTP requests
PyYAML>=6.0                # YAML configuration files

# Data Analysis
pandas>=2.2.0              # Data manipulation
numpy>=2.0.0               # Numerical computations
scipy>=1.11.0              # Scientific computing

# Optional - Visualization & Dashboard
plotly>=5.18.0             # Interactive charts
gradio>=4.0.0              # Web interface
streamlit>=1.29.0          # Dashboard (optional)
matplotlib>=3.8.0          # Plotting (optional)

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
freezegun>=1.2.0